% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.gene.updated.mouse.R
\docType{data}
\name{cc.gene.updated.mouse}
\alias{cc.gene.updated.mouse}
\title{Mouse cell cycle genes}
\format{
\subsection{\code{cc.gene.updated.mouse}}{

A list containing two character vectors of length 43 (S genes) and 54 (G2M genes), respectively.
}
}
\source{
TODO
}
\usage{
cc.gene.updated.mouse
}
\description{
A data object containing mouse cell cycle genes
}
\keyword{datasets}
