% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{create_pseudobulk_counts}
\alias{create_pseudobulk_counts}
\title{Create pseudobulk counts}
\usage{
create_pseudobulk_counts(sc_obj, log_flag)
}
\arguments{
\item{sc_obj}{Seurat object Input list of genes}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\description{
Create pseudobulk counts
}
\examples{
\dontrun{pseudobulk_counts <- create_pseudobulk_counts(sc_obj)}
}
