% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{Create_QC_Output_Prefiltered_ATAC}
\alias{Create_QC_Output_Prefiltered_ATAC}
\title{Create QC output files for prefiltered data (ATAC)}
\usage{
Create_QC_Output_Prefiltered_ATAC(proj, log_flag = FALSE)
}
\arguments{
\item{proj}{ArchR project associated with data}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
TRUE
}
\description{
Create QC output files for prefiltered data (ATAC)
}
\examples{
\dontrun{Create_QC_Output_Prefiltered_ATAC(proj = proj}
}
