% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{Create_QC_Output_Prefiltered_RNA}
\alias{Create_QC_Output_Prefiltered_RNA}
\title{Create QC output files for prefiltered data (RNA)}
\usage{
Create_QC_Output_Prefiltered_RNA(
  sc_obj,
  output_dir = getwd(),
  log_flag = FALSE
)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples (pre-filtering)}

\item{output_dir}{Directory where QC output will be saved}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
TRUE
}
\description{
Create QC output files for prefiltered data (RNA)
}
\examples{
\dontrun{Create_QC_Output_Prefiltered_RNA(sc_obj = sc_obj, output_dir = getwd()}
}
