% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc.R
\name{Create_RNA_QC_Output}
\alias{Create_RNA_QC_Output}
\title{Filter raw data for RNA}
\usage{
Create_RNA_QC_Output(sc_obj, output_dir = getwd(), log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples (pre-filtering)}

\item{output_dir}{Directory where QC output will be saved}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A Seurat object which contains filtered data from all input data
}
\description{
Filter raw data for RNA
}
\examples{
\dontrun{sc_obj <- Create_RNA_QC_Output(sc_obj = sc_obj, output_dir = getwd()}
}
