% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_SPEEDI_log_file}
\alias{create_SPEEDI_log_file}
\title{Creates log file for SPEEDI}
\usage{
create_SPEEDI_log_file(output_dir = getwd(), log_file_name = NULL)
}
\arguments{
\item{output_dir}{Output directory}

\item{log_file_name}{Log file name}
}
\value{
Open log file
}
\description{
Creates log file for SPEEDI
}
