% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_data.R
\name{FilterRawData}
\alias{FilterRawData}
\title{Filter raw data}
\usage{
FilterRawData(all_sc_exp_matrices, human)
}
\arguments{
\item{all_sc_exp_matrices}{list of single cell expression matrices}

\item{human}{flag to indicate whether we're processing human or mouse data}
}
\value{
A Seurat object which contains filtered data from all input data
}
\description{
Filter raw data
}
