% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_data.R
\name{FilterRawData_RNA}
\alias{FilterRawData_RNA}
\title{Filter raw data for RNA}
\usage{
FilterRawData_RNA(
  all_sc_exp_matrices,
  species = "human",
  record_doublets = FALSE,
  output_dir = getwd(),
  exit_with_code = FALSE,
  log_file_path = NULL,
  log_flag = FALSE
)
}
\arguments{
\item{all_sc_exp_matrices}{List of single cell expression matrices}

\item{species}{Species being analyzed. Possible choices are \code{"human"} or \code{"mouse"}.}

\item{record_doublets}{Boolean flag to indicate whether we will record doublets in the data (using the \link{scDblFinder} package). Possible choices are \code{TRUE} or \code{FALSE}.}

\item{output_dir}{Path to directory where output will be saved. Defaults to working directory (\code{\link[=getwd]{getwd()}}).}

\item{exit_with_code}{Boolean flag to indicate whether we will terminate R session with exit code (via \code{\link[=quit]{quit()}}) if error occurs. If set to FALSE, we just use \code{\link[=stop]{stop()}}.}

\item{log_file_path}{Path to log file (used to capture QC thresholds during parallel processing). Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A Seurat object which contains filtered data from all input data
}
\description{
Filter raw data for RNA
}
\examples{
\dontrun{sc_obj <- FilterRawData_RNA(all_sc_exp_matrices)}
\dontrun{sc_obj <- FilterRawData_RNA(all_sc_exp_matrices,
species = "human", record_doublets = TRUE)}
}
