% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_clusters_SPEEDI}
\alias{find_clusters_SPEEDI}
\title{Use either Louvain or Leiden for clustering. If using Leiden, tries Leiden first and then if it fails, uses Louvain}
\usage{
find_clusters_SPEEDI(sc_obj, resolution, method = "Louvain", log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{resolution}{Resolution for clustering}

\item{method}{Clustering method}

\item{log_flag}{boolean to indicate whether we're also printing to log file}
}
\value{
Seurat object with clustering complete
}
\description{
Use either Louvain or Leiden for clustering. If using Leiden, tries Leiden first and then if it fails, uses Louvain
}
