% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{FindMappingAnchors}
\alias{FindMappingAnchors}
\title{Find mapping anchors between reference and query}
\usage{
FindMappingAnchors(sc_obj, reference, data_type = "scRNA", log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{reference}{A Seurat reference object}

\item{data_type}{string to indicate whether we're analyzing scRNA data or other data}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
Mapping anchors between reference and query
}
\description{
Find mapping anchors between reference and query
}
\examples{
anchors <- FindMappingAnchors(sc_obj, reference = custom_reference_seurat_object, data_type = "scRNA")
}
