% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Compute geometric mean}
\usage{
geometric.mean(x, y)
}
\arguments{
\item{x}{First number}

\item{y}{Second number}
}
\value{
Geometric mean of two numbers
}
\description{
Compute geometric mean
}
