% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{grab_hb_networks}
\alias{grab_hb_networks}
\title{Grab associated HumanBase networks for current cell type and reference tissue}
\usage{
grab_hb_networks(current_cell_type, reference_tissue, log_flag = FALSE)
}
\arguments{
\item{current_cell_type}{Current cell type}

\item{reference_tissue}{Reference tissue}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A vector containing a list of HumanBase networks
}
\description{
Grab associated HumanBase networks for current cell type and reference tissue
}
\examples{
\dontrun{hb_networks <- grab_hb_networks(current_cell_type = "B naive",
reference_tissue = "pbmc")}
}
