% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batches.R
\name{InferBatches}
\alias{InferBatches}
\title{Infer batches using LISI metric}
\usage{
InferBatches(sc_obj)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}
}
\value{
A Seurat object which contains labeled batches
}
\description{
Infer batches using LISI metric
}
