% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_data.R
\name{InitialProcessing}
\alias{InitialProcessing}
\title{Process filtered data}
\usage{
InitialProcessing(sc_obj, species = "human", log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{species}{flag to indicate whether we're processing human or mouse data}

\item{log_flag}{if set to TRUE, we previously set up a log file where certain output will be written (e.g., parameters)}
}
\value{
A Seurat object which contains processed data
}
\description{
Process filtered data
}
