% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batches.R
\name{IntegrateByBatch}
\alias{IntegrateByBatch}
\title{Integrate batches}
\usage{
IntegrateByBatch(sc_obj)
}
\arguments{
\item{sc_obj}{Seurat object}
}
\value{
A Seurat object which contains integrated data
}
\description{
Integrate batches
}
