% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batches.R
\name{IntegrateByBatch_ATAC}
\alias{IntegrateByBatch_ATAC}
\title{Integrate batches (ATAC)}
\usage{
IntegrateByBatch_ATAC(proj, log_flag = FALSE)
}
\arguments{
\item{proj}{ArchR project containing cells for all samples}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
An ArchR object which contains integrated data
}
\description{
Integrate batches (ATAC)
}
\examples{
\dontrun{proj <- IntegrateByBatch_ATAC(proj)}
}
