% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batches.R
\name{IntegrateByBatch_RNA}
\alias{IntegrateByBatch_RNA}
\title{Integrate batches (RNA)}
\usage{
IntegrateByBatch_RNA(sc_obj, log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A Seurat object which contains integrated data
}
\description{
Integrate batches (RNA)
}
\examples{
sc_obj <- IntegrateByBatch_RNA(sc_obj)
}
