% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lappend}
\alias{lappend}
\title{Append a list to a list-of-lists}
\usage{
lappend(lst, ...)
}
\arguments{
\item{lst}{List}

\item{...}{Additional lists}
}
\value{
A list of lists
}
\description{
Append a list to a list-of-lists
}
