% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{MajorityVote_ATAC}
\alias{MajorityVote_ATAC}
\title{In each cluster, we vote for majority cell type (ATAC)}
\usage{
MajorityVote_ATAC(proj, log_flag = FALSE)
}
\arguments{
\item{proj}{ArchR object containing cells for all samples}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
An ArchR object which contains majority vote labels
}
\description{
In each cluster, we vote for majority cell type (ATAC)
}
\examples{
\dontrun{proj <- MajorityVote_ATAC(proj)}
}
