% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{MapCellTypes}
\alias{MapCellTypes}
\title{Map cell types for input data}
\usage{
MapCellTypes(sc_obj, reference, data_type = "scRNA")
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{reference}{Seurat reference object}

\item{data_type}{string to indicate whether we're analyzing scRNA or snRNA data}
}
\value{
A Seurat object which contains majority vote labels
}
\description{
Map cell types for input data
}
