% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{MapCellTypes_ATAC}
\alias{MapCellTypes_ATAC}
\title{Map cell types for input data (ATAC)}
\usage{
MapCellTypes_ATAC(
  proj,
  reference,
  reference_cell_type_attribute = "celltype.l2",
  exit_with_code = FALSE,
  log_flag = FALSE
)
}
\arguments{
\item{proj}{ArchR project containing cells for all samples}

\item{reference}{Seurat reference object}

\item{reference_cell_type_attribute}{If using a Seurat reference object, this parameter captures where the cell type information is stored}

\item{exit_with_code}{Boolean flag to indicate whether we will terminate R session with exit code (via \code{\link[=quit]{quit()}}) if error occurs. If set to FALSE, we just use \code{\link[=stop]{stop()}}.}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
An ArchR object which contains majority vote labels
}
\description{
Map cell types for input data (ATAC)
}
\examples{
\dontrun{sc_obj <- MapCellTypes_ATAC(proj, reference = custom_reference_seurat_object)}
}
