% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{MapCellTypes_RNA}
\alias{MapCellTypes_RNA}
\title{Map cell types for input data (RNA)}
\usage{
MapCellTypes_RNA(sc_obj, reference, data_type = "scRNA", log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{reference}{Seurat reference object or reference found in \link{SeuratData}}

\item{data_type}{String to indicate whether we're analyzing scRNA or snRNA data}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A Seurat object which contains majority vote labels
}
\description{
Map cell types for input data (RNA)
}
\examples{
sc_obj <- MapCellTypes_RNA(sc_obj, reference = custom_reference_seurat_object)
sc_obj <- MapCellTypes_RNA(sc_obj, reference = "adiposeref")
}
