% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_dir_path}
\alias{normalize_dir_path}
\title{Normalize directory path for processing}
\usage{
normalize_dir_path(path)
}
\arguments{
\item{path}{Directory path}
}
\value{
Normalized directory path
}
\description{
Normalize directory path for processing
}
