% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_namespace_and_function}
\alias{parse_namespace_and_function}
\title{Parse namespace and function name for do.call}
\usage{
parse_namespace_and_function(x)
}
\arguments{
\item{x}{Function name (potentially with namespace attached)}
}
\value{
Parsed function
}
\description{
Parse namespace and function name for do.call
}
