% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preliminary_check_for_SPEEDI_errors}
\alias{preliminary_check_for_SPEEDI_errors}
\title{Preliminary check for SPEEDI errors}
\usage{
preliminary_check_for_SPEEDI_errors(
  data_type,
  reference_tissue,
  log_flag = FALSE
)
}
\arguments{
\item{data_type}{Data type}

\item{reference_tissue}{Reference tissue}

\item{log_flag}{boolean to indicate whether we're also printing to log file}
}
\value{
exit code indicating error
}
\description{
Preliminary check for SPEEDI errors
}
