% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PrepDR}
\alias{PrepDR}
\title{Do stuff}
\usage{
PrepDR(object, features = NULL, slot = "scale.data", verbose = TRUE)
}
\arguments{
\item{object}{stuff}

\item{features}{features}

\item{slot}{slot to use}

\item{verbose}{whether to give detailed feedback}
}
\value{
stuff
}
\description{
Do stuff
}
