% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_heatmap_cell_type_proportions_RNA}
\alias{print_heatmap_cell_type_proportions_RNA}
\title{Print heatmap for cell type proportions in Seurat object (RNA)}
\usage{
print_heatmap_cell_type_proportions_RNA(
  sc_obj,
  file_name,
  output_dir = getwd(),
  log_flag = FALSE
)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{file_name}{File name for saved plot}

\item{output_dir}{Path to directory where output will be saved. Defaults to working directory (\code{\link[=getwd]{getwd()}}).}

\item{log_flag}{boolean to indicate whether we're also printing to log file}
}
\value{
TRUE
}
\description{
Print heatmap for cell type proportions in Seurat object (RNA)
}
