% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_UMAP}
\alias{print_UMAP}
\title{Print UMAP for Seurat object}
\usage{
print_UMAP(
  sc_obj,
  file_name,
  group_by_category = NULL,
  output_dir = getwd(),
  log_flag = FALSE
)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{file_name}{File name for saved plot}

\item{group_by_category}{Category to group samples by}

\item{output_dir}{Path to directory where output will be saved. Defaults to working directory (\code{\link[=getwd]{getwd()}}).}

\item{log_flag}{boolean to indicate whether we're also printing to log file}
}
\value{
TRUE
}
\description{
Print UMAP for Seurat object
}
