% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quit_SPEEDI}
\alias{quit_SPEEDI}
\title{Quits SPEEDI (either with exit code if exit_with_code == TRUE or with stop())}
\usage{
quit_SPEEDI(exit_with_code, exit_code = 0, log_flag = FALSE)
}
\arguments{
\item{exit_with_code}{Boolean flag to indicate whether we will terminate R session with exit code (via \code{\link[=quit]{quit()}}) if error occurs. If set to FALSE, we just use \code{\link[=stop]{stop()}}.}

\item{exit_code}{exit code}

\item{log_flag}{Boolean to indicate whether we're also printing to log file}
}
\value{
TRUE
}
\description{
Quits SPEEDI (either with exit code if exit_with_code == TRUE or with stop())
}
