% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{Read_ATAC}
\alias{Read_ATAC}
\title{Read in ATAC data for processing}
\usage{
Read_ATAC(
  data_path = getwd(),
  output_dir = getwd(),
  sample_id_list = NULL,
  sample_file_paths = NULL,
  species = "human",
  log_flag = FALSE
)
}
\arguments{
\item{data_path}{Path to directory where input data are located. Defaults to working directory (\code{\link[=getwd]{getwd()}}).}

\item{output_dir}{Path to directory where output will be saved. Defaults to working directory (\code{\link[=getwd]{getwd()}}).}

\item{sample_id_list}{Vector of sample names (optional - if not provided, will select all samples found recursively in \code{data_path}).}

\item{sample_file_paths}{Vector of sample file paths (optional - if not provided, will select all samples found recursively in \code{data_path}). If used, \code{sample_id_list} is required.}

\item{species}{Species being analyzed. Possible choices are \code{"human"} or \code{"mouse"}.}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
An ArchR project with associated Arrow files
}
\description{
Read in ATAC data for processing
}
\examples{
\dontrun{proj <- Read_ATAC()}
\dontrun{proj <- Read_ATAC(data_path = "~/input_data/",
sample_id_list = c("sample_1", "sample_2"), species = "human")}
}
