% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{Read_h5}
\alias{Read_h5}
\title{Read in h5 data for processing}
\usage{
Read_h5(data_path = getwd(), sample_id_list = NULL, log_flag = FALSE)
}
\arguments{
\item{data_path}{Path to directory where input data are located. Defaults to working directory (\code{\link[=getwd]{getwd()}}).}

\item{sample_id_list}{Vector of sample names (optional - if not provided, will select all samples found recursively in \code{data_path}).}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A set of single cell expression matrices
}
\description{
Read in h5 data for processing
}
\examples{
all_sc_exp_matrices <- Read_h5()
all_sc_exp_matrices <- Read_h5(data_path = "~/input_data/", sample_id_list = c("sample_1", "sample_2"))
}
