% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RowSparseCheck}
\alias{RowSparseCheck}
\title{Check if the input matrix is dgCMatrix}
\usage{
RowSparseCheck(mat)
}
\arguments{
\item{mat}{sparse matrix}
}
\value{
A dgCMatrix
}
\description{
Check if the input matrix is dgCMatrix
}
