% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RowVar}
\alias{RowVar}
\title{Calculate row variance of a dense matrix}
\usage{
RowVar(mat)
}
\arguments{
\item{mat}{dense matrix}
}
\value{
A vector of row variance
}
\description{
Calculate row variance of a dense matrix
}
