% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RowVarSparse}
\alias{RowVarSparse}
\title{Calculate row variance of a sparse matrix}
\usage{
RowVarSparse(mat)
}
\arguments{
\item{mat}{sparse matrix}
}
\value{
A vector of row variance
}
\description{
Calculate row variance of a sparse matrix
}
