% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run_SPEEDI}
\alias{run_SPEEDI}
\title{Wrapper function for running the SPEEDI pipeline}
\usage{
run_SPEEDI(
  data_path,
  sample_id_list,
  tissue,
  human = TRUE,
  remove_doublets = FALSE
)
}
\arguments{
\item{data_path}{path to where data is located}

\item{sample_id_list}{list of sample names (optional - if not provided, will select all samples found recursively in data_path)}

\item{tissue}{Tissue of data (used to map cell types via reference)}

\item{human}{flag to indicate whether we're processing human or mouse data}

\item{remove_doublets}{flag to indicate whether we're removing doublets from the data (using scDblFinder)}
}
\value{
A Seurat object that has been processed through the SPEEDI pipeline
}
\description{
Wrapper function for running the SPEEDI pipeline
}
