% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{RunDE_RNA}
\alias{RunDE_RNA}
\title{Perform differential expression analysis (RNA)}
\usage{
RunDE_RNA(sc_obj, metadata_df, output_dir = getwd(), log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{metadata_df}{Dataframe containing metadata for samples. Rownames should be sample names and column names should be metadata attributes with two classes (e.g., condition: disease and control)}

\item{output_dir}{Path to directory where output will be saved. Defaults to working directory (\code{\link[=getwd]{getwd()}}). Directory will be created if it doesn't already exist.}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A list containing differential expression analyses
}
\description{
Perform differential expression analysis (RNA)
}
\examples{
\dontrun{differential_expression_results <- RunDE_RNA(sc_obj = sc_obj, metadata_df = metadata_df)}
}
