% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{RunFMD_RNA}
\alias{RunFMD_RNA}
\title{Perform functional module discovery (RNA)}
\usage{
RunFMD_RNA(gene_list, network = "global", log_flag = FALSE)
}
\arguments{
\item{gene_list}{List of genes}

\item{network}{Background network}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A list (first element containing URL to see full results in web browser and second element containing enrichment results)
}
\description{
Perform functional module discovery (RNA)
}
\examples{
\dontrun{fmd_results <- RunFMD_RNA(gene_list = c("ADNP", "CHD8"), network = "blood")}
}
