% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_zero_one}
\alias{scale_zero_one}
\title{Scale a vector to range(0,1)}
\usage{
scale_zero_one(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A scaled vector ranging from 0 to 1
}
\description{
Scale a vector to range(0,1)
}
