% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_pbmc_reference_url}
\alias{set_pbmc_reference_url}
\title{Set PBMC reference URL}
\usage{
set_pbmc_reference_url(
  new_url = "https://atlas.fredhutch.org/data/nygc/multimodal/pbmc_multimodal.h5seurat"
)
}
\arguments{
\item{new_url}{the new PBMC reference url}
}
\description{
Set PBMC reference URL
}
