% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_speedi_seed}
\alias{set_speedi_seed}
\title{Set SPEEDI random seed}
\usage{
set_speedi_seed(new_seed = 1253)
}
\arguments{
\item{new_seed}{the new SPEEDI random seed}
}
\description{
Set SPEEDI random seed
}
