% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{SetDefaultAssay}
\alias{SetDefaultAssay}
\title{Choose assay based on whether there are multiple batches (integrated) or only one batch (SCT)}
\usage{
SetDefaultAssay(sc_obj)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}
}
\value{
A Seurat object with default assay appropriately set
}
\description{
Choose assay based on whether there are multiple batches (integrated) or only one batch (SCT)
}
