% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Stops SPEEDI quietly (so we don't write superfluous error messages)}
\usage{
stop_quietly()
}
\value{
TRUE
}
\description{
Stops SPEEDI quietly (so we don't write superfluous error messages)
}
