% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_multiome.R
\name{TransferRNALabels}
\alias{TransferRNALabels}
\title{Transfer cell type labels from RNA to ATAC (only works for true multiome since same cells in both assays)}
\usage{
TransferRNALabels(sc_obj, proj, exit_with_code = FALSE, log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{proj}{ArchR project associated with data}

\item{exit_with_code}{Boolean flag to indicate whether we will terminate R session with exit code (via \code{\link[=quit]{quit()}}) if error occurs. If set to FALSE, we just use \code{\link[=stop]{stop()}}.}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
An ArchR project with cell type labels transferred from RNA
}
\description{
Transfer cell type labels from RNA to ATAC (only works for true multiome since same cells in both assays)
}
\examples{
\dontrun{sc_obj <- FindMultiomeOverlap(sc_obj, proj, data_modality = "RNA")}
\dontrun{proj <- FindMultiomeOverlap(sc_obj, proj, data_modality = "ATAC")}
}
