% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_data.R
\name{FilterRawData}
\alias{FilterRawData}
\title{Filter raw data}
\usage{
FilterRawData(
  all_sc_exp_matrices,
  species = "human",
  record_doublets = FALSE,
  log_file_path = NULL,
  log_flag = FALSE
)
}
\arguments{
\item{all_sc_exp_matrices}{list of single cell expression matrices}

\item{species}{flag to indicate whether we're processing human or mouse data}

\item{record_doublets}{flag to indicate whether we're recording doublet info in the data (using scDblFinder)}

\item{log_file_path}{path to log file (used to capture QC thresholds during parallel processing)}

\item{log_flag}{if set to TRUE, we previously set up a log file where certain output will be written (e.g., parameters)}
}
\value{
A Seurat object which contains filtered data from all input data
}
\description{
Filter raw data
}
