% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_clusters_SPEEDI}
\alias{find_clusters_SPEEDI}
\title{First, try Leiden algorithm for clustering. If that doesn't work, then try Louvain}
\usage{
find_clusters_SPEEDI(sc_obj, resolution, log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{resolution}{Resolution for clustering}

\item{log_flag}{boolean to indicate whether we're also printing to log file}
}
\value{
Seurat object with clustering complete
}
\description{
First, try Leiden algorithm for clustering. If that doesn't work, then try Louvain
}
