% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{FindMappingAnchors}
\alias{FindMappingAnchors}
\title{Find mapping anchors between reference and query}
\usage{
FindMappingAnchors(sc_obj, reference)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{reference}{A Seurat reference object}
}
\value{
Mapping anchors between reference and query
}
\description{
Find mapping anchors between reference and query
}
