% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{FindMappingAnchors}
\alias{FindMappingAnchors}
\title{Find mapping anchors between reference and query}
\usage{
FindMappingAnchors(sc_obj, reference, data_type = "scRNA")
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{reference}{A Seurat reference object}

\item{data_type}{string to indicate whether we're analyzing scRNA or snRNA data}
}
\value{
Mapping anchors between reference and query
}
\description{
Find mapping anchors between reference and query
}
