% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_multiome.R
\name{FindMultiomeOverlap}
\alias{FindMultiomeOverlap}
\title{Find overlap between true multiome cells (RNA / ATAC)}
\usage{
FindMultiomeOverlap(
  sc_obj,
  proj,
  data_modality = "RNA",
  output_dir = getwd(),
  exit_with_code = FALSE,
  log_flag = FALSE
)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{proj}{ArchR project associated with data}

\item{data_modality}{data modality to return (in Seurat object or ArchR project)}

\item{output_dir}{Path to directory where output will be saved. Defaults to working directory (\code{\link[=getwd]{getwd()}}).}

\item{exit_with_code}{Boolean flag to indicate whether we will terminate R session with exit code (via \code{\link[=quit]{quit()}}) if error occurs. If set to FALSE, we just use \code{\link[=stop]{stop()}}.}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A Seurat object or ArchR project with only those cells that pass QC in the other assay
}
\description{
Find overlap between true multiome cells (RNA / ATAC)
}
\examples{
\dontrun{sc_obj <- FindMultiomeOverlap(sc_obj, proj, data_modality = "RNA")}
\dontrun{proj <- FindMultiomeOverlap(sc_obj, proj, data_modality = "ATAC")}
}
