% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batches.R
\name{InferBatches}
\alias{InferBatches}
\title{Infer batches using LISI metric}
\usage{
InferBatches(sc_obj, log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A Seurat object which contains labeled batches
}
\description{
Infer batches using LISI metric
}
\examples{
\dontrun{sc_obj <- InferBatches(sc_obj)}
}
