% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batches.R
\name{InferBatches}
\alias{InferBatches}
\title{Infer batches using LISI metric}
\usage{
InferBatches(sc_obj, log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{log_flag}{if set to TRUE, we previously set up a log file where certain output will be written (e.g., parameters)}
}
\value{
A Seurat object which contains labeled batches
}
\description{
Infer batches using LISI metric
}
