% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_data.R
\name{InitialProcessing_ATAC}
\alias{InitialProcessing_ATAC}
\title{Process filtered data (ATAC)}
\usage{
InitialProcessing_ATAC(proj, log_flag = FALSE)
}
\arguments{
\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}

\item{sc_obj}{Seurat object containing cells for all samples}
}
\value{
An ArchR project which contains processed data
}
\description{
Process filtered data (ATAC)
}
\examples{
sc_obj <- InitialProcessing_RNA(sc_obj)
}
