% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_data.R
\name{InitialProcessing_RNA}
\alias{InitialProcessing_RNA}
\title{Process filtered data (RNA)}
\usage{
InitialProcessing_RNA(sc_obj, species = "human", log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{species}{Species being analyzed. Possible choices are \code{"human"} or \code{"mouse"}.}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A Seurat object which contains processed data
}
\description{
Process filtered data (RNA)
}
\examples{
sc_obj <- InitialProcessing_RNA(sc_obj)
sc_obj <- InitialProcessing_RNA(sc_obj, species = "human")
}
