% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batches.R
\name{IntegrateByBatch}
\alias{IntegrateByBatch}
\title{Integrate batches}
\usage{
IntegrateByBatch(sc_obj, log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object}

\item{log_flag}{if set to TRUE, we previously set up a log file where certain output will be written (e.g., parameters)}
}
\value{
A Seurat object which contains integrated data
}
\description{
Integrate batches
}
