% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{LoadReference}
\alias{LoadReference}
\title{Load appropriate reference}
\usage{
LoadReference(tissue, human, reference_path = getwd())
}
\arguments{
\item{tissue}{tissue associated with samples}

\item{human}{flag to indicate whether sample is human or mouse}
}
\value{
A reference object
}
\description{
Load appropriate reference
}
