% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{LoadReferenceSPEEDI}
\alias{LoadReferenceSPEEDI}
\title{Load appropriate reference for SPEEDI}
\usage{
LoadReferenceSPEEDI(
  tissue,
  species = "human",
  reference_dir = getwd(),
  reference_file_name = NULL,
  log_flag = FALSE
)
}
\arguments{
\item{tissue}{tissue associated with samples}

\item{species}{flag to indicate whether sample is human or mouse}

\item{reference_dir}{path to base directory for reference}

\item{reference_file_name}{name of reference file, inside of reference_dir, that should be loaded (only used if tissue is set to custom)}

\item{log_flag}{if set to TRUE, we previously set up a log file where certain output will be written (e.g., parameters)}
}
\value{
A reference object
}
\description{
Load appropriate reference for SPEEDI
}
