% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{LoadReferenceSPEEDI}
\alias{LoadReferenceSPEEDI}
\title{Load appropriate reference - need to update function name so it doesn't overlap with Azimuth}
\usage{
LoadReferenceSPEEDI(tissue, human, reference_path = getwd())
}
\arguments{
\item{tissue}{tissue associated with samples}

\item{human}{flag to indicate whether sample is human or mouse}

\item{reference_path}{path to base directory for reference (or maybe full path to reference?)}
}
\value{
A reference object
}
\description{
Load appropriate reference - need to update function name so it doesn't overlap with Azimuth
}
