% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{LoadReferenceSPEEDI}
\alias{LoadReferenceSPEEDI}
\title{Load appropriate reference for SPEEDI}
\usage{
LoadReferenceSPEEDI(
  tissue,
  species = "human",
  reference_dir = getwd(),
  reference_file_name = NULL
)
}
\arguments{
\item{tissue}{tissue associated with samples}

\item{species}{flag to indicate whether sample is human or mouse}

\item{reference_dir}{path to base directory for reference}

\item{reference_file_name}{name of reference file, inside of reference_dir, that should be loaded (only used if tissue is set to custom)}
}
\value{
A reference object
}
\description{
Load appropriate reference for SPEEDI
}
