% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{MajorityVote}
\alias{MajorityVote}
\title{In each cluster, vote for majority cell type}
\usage{
MajorityVote(sc_obj, used_assay, current_resolution = 1)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{used_assay}{assay used for processing (integrated or SCT)}

\item{current_resolution}{parameter that indicates current resolution for clustering}
}
\value{
A Seurat object which contains majority vote labels
}
\description{
In each cluster, vote for majority cell type
}
