% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{Read_h5}
\alias{Read_h5}
\title{Read in data}
\usage{
Read_h5(data_path, sample_id_list)
}
\arguments{
\item{data_path}{path to where data is located}

\item{sample_id_list}{list of sample names}
}
\value{
A set of single cell expression matrices
}
\description{
Read in data
}
