% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run_SPEEDI}
\alias{run_SPEEDI}
\title{Wrapper function for running the SPEEDI pipeline}
\usage{
run_SPEEDI(
  tissue,
  data_path = getwd(),
  reference_dir = getwd(),
  output_dir = getwd(),
  reference_file_name = NULL,
  sample_id_list = NULL,
  species = "human",
  record_doublets = FALSE
)
}
\arguments{
\item{tissue}{tissue of data (used to map cell types via reference)}

\item{data_path}{path to where data is located}

\item{reference_dir}{path to dir where reference is located}

\item{output_dir}{path to dir where output will be saved}

\item{sample_id_list}{list of sample names (optional - if not provided, will select all samples found recursively in data_path)}

\item{species}{flag to indicate whether we're processing human or mouse data}

\item{record_doublets}{flag to indicate whether we're recording doublets in the data (using scDblFinder)}
}
\value{
A Seurat object that has been processed through the SPEEDI pipeline
}
\description{
Wrapper function for running the SPEEDI pipeline
}
