% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{SetDefaultAssay}
\alias{SetDefaultAssay}
\title{Choose assay based on whether there are multiple batches (integrated) or only one batch (SCT)}
\usage{
SetDefaultAssay(sc_obj, log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A Seurat object with default assay appropriately set
}
\description{
Choose assay based on whether there are multiple batches (integrated) or only one batch (SCT)
}
\examples{
\dontrun{sc_obj <- SetDefaultAssay(sc_obj)}
}
