% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{SetDefaultAssay}
\alias{SetDefaultAssay}
\title{Choose assay based on whether there are multiple batches (integrated) or only one batch (SCT)}
\usage{
SetDefaultAssay(sc_obj, log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{log_flag}{if set to TRUE, we previously set up a log file where certain output will be written (e.g., parameters)}
}
\value{
A Seurat object with default assay appropriately set
}
\description{
Choose assay based on whether there are multiple batches (integrated) or only one batch (SCT)
}
