% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{SetPredictedId}
\alias{SetPredictedId}
\title{We always want to use the predicted.id column in our Seurat object to determine majority vote.
However, if we use Azimuth::RunAzimuth, there are often predictions made at multiple annotation levels.
Depending on the reference, we use SetPredictedId to select the proper annotation level for predicted.id.}
\usage{
SetPredictedId(sc_obj, reference)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{reference}{Reference name}
}
\value{
A Seurat object with predicted.id appropriately set
}
\description{
We always want to use the predicted.id column in our Seurat object to determine majority vote.
However, if we use Azimuth::RunAzimuth, there are often predictions made at multiple annotation levels.
Depending on the reference, we use SetPredictedId to select the proper annotation level for predicted.id.
}
