% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cell_type_from_reference.R
\name{SetPredictedId}
\alias{SetPredictedId}
\title{We always want to use the \code{predicted.id} column in our Seurat object to determine majority vote.
However, if we use \link[Azimuth:RunAzimuth]{Azimuth::RunAzimuth}, there are often predictions made at multiple annotation levels.
Depending on the reference, we use \code{SetPredictedId} to select the proper annotation level for predicted.id.}
\usage{
SetPredictedId(sc_obj, reference, log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{reference}{Reference name}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A Seurat object with predicted.id appropriately set
}
\description{
We always want to use the \code{predicted.id} column in our Seurat object to determine majority vote.
However, if we use \link[Azimuth:RunAzimuth]{Azimuth::RunAzimuth}, there are often predictions made at multiple annotation levels.
Depending on the reference, we use \code{SetPredictedId} to select the proper annotation level for predicted.id.
}
\examples{
sc_obj <- SetPredictedId(sc_obj, reference = "bonemarrowref")
}
