% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batches.R
\name{VisualizeIntegration}
\alias{VisualizeIntegration}
\title{Visualize integration and prepare SCT for finding markers}
\usage{
VisualizeIntegration(sc_obj)
}
\arguments{
\item{sc_obj}{Seurat object}
}
\value{
A Seurat object with SCT markers and visualizations
}
\description{
Visualize integration and prepare SCT for finding markers
}
