% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batches.R
\name{VisualizeIntegration}
\alias{VisualizeIntegration}
\title{Visualize integration and prepare SCT for finding markers}
\usage{
VisualizeIntegration(sc_obj, exit_with_code = FALSE, log_flag = FALSE)
}
\arguments{
\item{sc_obj}{Seurat object containing cells for all samples}

\item{exit_with_code}{Boolean flag to indicate whether we will terminate R session with exit code (via \code{\link[=quit]{quit()}}) if error occurs. If set to FALSE, we just use \code{\link[=stop]{stop()}}.}

\item{log_flag}{If set to TRUE, record certain output (e.g., parameters) to a previously set up log file. Most likely only used in the context of \code{\link[=run_SPEEDI]{run_SPEEDI()}}.}
}
\value{
A Seurat object with SCT markers and visualizations
}
\description{
Visualize integration and prepare SCT for finding markers
}
\examples{
\dontrun{sc_obj <- VisualizeIntegration(sc_obj)}
}
